+++ test/integration/user/test/MVT.test.js
@@ -1,4 +1,4 @@
-import * as carto from '../../../../src/';
+import carto from '../../../../src/';
 import * as util from '../../util';
 
 describe('Layer', () => {